Moved namespaces for importer service. Provider and tour repos
