+++ ApiIntegration/Services/Importer.cs
@@ -3,7 +3,7 @@ using Microsoft.Extensions.Logging;
 using System;
 using System.Threading.Tasks;
 
-namespace ApiIntegration
+namespace ApiIntegration.Services
 {
     public class Importer : IImporter
     {
